# optimize and cleanup